backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
chromadb==0.5.20
voyageai==0.3.2
anthropic==0.40.0
pdfplumber==0.11.4
python-dotenv==1.0.1
pydantic==2.10.0
